Improve performance of InertialProvider(Frame)
It now has better performance when the attitude is aligned with the
propagation reference frame by avoiding multiplying by zero. With
analytic propagators using this attitude provider instead of the default
has been observed to result in a speed up of 2 in some cases, e.g. with
the TLEPropagator.

For #586