Align attitude with propagation frame
Implemented for all analytic and the GlONASS propagator. The DSST and
Numerical propagator do not have a frame available in their constructor,
so the new behavior could not be implemented. Aligning the attitude
frame with the propagation frame speeds up propagation when the user
does not care about attitude and it many cases it removes a dependence
on the default data context.

Fix #586