Fix step handled twice in analytic with event
Previously when an analytic propagator handled a continue event the
first part of the step would be passed to the OrekitStepHandler twice.
Now it is only passed once. Seems to have been present when
AbstractAnalyticalPropagator was created, about a decade ago, and not
detected in all the changes since then, so I guess no one uses that
combination much.