Adding truncated SOLFSMY test file / Tests correction
The first tests were designed for the full solar indices file, while now
it has been adapted for a truncated version.
A bug has been spotted in the division of the lines, with values being
limited to 99.9. This will be avoided by using String.split(),
considering all values to separated by spaces.
A bug was spotted on the propagation test that might be due to the
ephemeris size, leading to the choice of the truncated dates (2004 to
2007).