JB2008 Model : DSTDTC parsing added
Minor Modifications :
- The JB2008SpaceWeatherData has been renamed JB2008SpaceEnvironmentData
- The pattern matching used to get information from lines has been
transformed to a static version, compiled outside of the lines loop.
- Lags were added on the different parameters, through the
interpolation, and modifying the definition of the first and last dates
accessible.

Major Modifications :
- A DtcDataLoader has been coded based on the SOLFSMYDataLoader.
- The JB2008SpaceEnvironmentData class has been modified not to inherit
from AbstractSelfFeedingLoader, to allow for the use of 2 loaders in the
same class. Additional parameters and methods (ending by DTC) were added
to facilitate the implementation of the second loader. A constructor was
added to include DTC information.
- Tests for the DtcDataLoader were implemented with the corresponding
test files.