Coverage and Duplication improvements for JB2008 DataLoaders
Duplicated lines :
- Most of the duplicated code was located in the LineReader object,
therefore, this code was moved outside of the DataLoaders in a common
class. The Deprecated flag was set on the CssiSpaceWeather dataloader
for implementation of the common class at the next major version.

Coverage :
- An unnecessary test was removed for loadData fonctions.
- Tests were set up for the CommonLineReader class, allowing the
validation for various dataloaders using this class.
- Tests were set up for the bracketDate fonctions.
- A test for the parsedEpochs check has been set, with the
implementation of an exception UNABLE_TO_PARSE_LINE_IN_FILE in case of a
double line. The exception might be adapted in the future.