Fix StreamingOemWriter in ITRF w/o optional fields
In Orekit 10.x StreamingOemWriter had the option for the user to select
the output reference frame and to not write optional metadata fields.
This commit adds those features to the 11.x branch while maintaining
compatibility with the existing 11.x releases.

A constructor parameter is added to StreamingOemWriter that determines
wheter the reference frame of each segment changes depending on the
state's attitude's frame or if the reference frame from the metadata is
used.

Interpolation degree values less than zero are treated as null when
writing an OEM.

Added a test case.