Fixed StateCovariance
Fixed shiftedBy method to consider cases where the state covariance is expressed in LOFType or non-pseudo inertial frames. Modified the OrbitType to EQUINOCTIAL to avoid singularities while keeping the same expected behavior.

Improved constructor to throw an error if input frame is not pseudo-inertial and given OrbitType is not CARTESIAN.

Minor fixes to documentation