Fixed #1300 : Regression in ephemeris interpolation
Fix interpolation in Ephemeris when interpolationPoints=1. I.e. stepwise
interpolations. Thanks to Sander Cochran for the failing test case.

Add back getCachedSamples() for #1300

Should improve backwards compatibility to make it suitable for a patch
release.

Deprecate SpacecraftStateInterpolator constructor

Deprecate a constructor that can lead to the SpacecraftStateInterpolator
being in an inconsistent state. For #1300

Fix #1300 for FieldInterpolator

Applied changes from non-field version to the field version.