Fix DSST Jacobians for force model parameters
Previously the DSST would throw an exception if attempting to compute
the jacobian w.r.t.  a force model parameter. Now it seems to work.

Changes include:
 * initializing field short period terms when computing Jacobians.
 * using EquinoctialOrbit, not CartesianOrbit for DSST.
 * only consider the short period terms for parameters in the force
   model that created the short period terms.

I added tests that check running the DSST with zero, one, or two
parameters in mean and osculating mode. Had to slightly increase the
tolerance on an existing DSST paramter Jacobian test. But the DSST
LAGEOS OD test completed in fewer iteration, leading me to believe it
was an improvement to the Jacobian.

I also addes some supporting test and debugging code.

Fixes #986