Fix DSST Kalman Filter tests #986
Previous change caused some of the DSST KF tests to no longer work.
Those tests seem to create a MEAN DSST propagator, then replace all of
the short period terms with OSCULATING ones. I'm not sure if this even
makes sense from the theory or should be supported, but apparently it
worked before.

The WLS OD test seems to converge faster now to the same solution, which
seems to indicate better derivatives.