Estimate SRP in DSST OD test GPS07 #986
At the request of Bryan Cazabonne the GPS07 test now estimates drag. The
min, max, and standard deviation of the residuals are all closer to
zero, which is consistent with estimating more parameters. It does take
significantly more iterations, with a similar number of evaluations.
Could indicate observations are not very sensitive to SRP, or SRP is
nearly linearly dependent on other parameters, or the solution becomes
sensitive to data editing. Probably does not indicate incorrect
derivatives because in that case Levenber Marquardt uses many more
evaluations than iterations. I don't know enought about the test case to
objectively state whether the changes are an improvement or not.