Skip writing OCM fields that are not present #1623
Previously it was impossible to tell from the Ocm returned by OcmParser
whether some optional fields with no defaults were present in the parsed
OCM or not. Similarly OcmWriter would write many optional fields without
default values whether they contained useful data or not.

Now optional fields without default values are initialized to null or
NaN so that it is possible to tell if they are present. Values that are
not present are not written.

Includes a test case for OCM. And some updates to existing test cases
that expected TRAJ_BASIS = PREDICTED even if that was never specified in
the input data.