Add option for CCSDS files to use 16 digits #1297
The CCSDS specs for NDM require that fixed and floating point numbers
contain no more than 16 digits in the mantissa. C.f. 502b3 S 7.5.6,
7.5.7.

Previously Orekit would use enough digits to guarentee round trip
safety, which can be 17 digits. Also some values less than 1 were
written in fixed point notation and the leading zeros caused the 16
limit to be exceeded.

Extracted a Formatter interface from the AccurateFormatter class,
updated the existing CCSDS writers to use an instance of Formatter, and
created a new implementation for the truncated CCSDS format.

The methods in Formatter are named "toString" in order to avoid a name
clash with the static "format" methods of AccurateFormatter, so that
backwards compatibility could be maintained.

Also added test cases for the new code and a test case that checks that
all CCSDS NDM write numbers with no more than 16 digits when using the
new formatter.

Thanks to Brad Kelly for reporting the bug and John Ajamian for
implementing the fix.