Adjust the number of CPUs required by the K8S pod running a CI job.
The CI/CD platform uses a K8S cluster that dynamically adjusts the number of
nodes to the load. To avoid over-consumption of resources, which would end up
costing the project a lot of money, the GitLab Runner is configured to require
few resources by default (2 cores, 4 GB RAM). These resources are sufficient
for most jobs. But the verify job can massively parallelize tasks, and runs
faster if it has more cores. In the CI/CD pipeline, we therefore adjust the
resources required by each job.