Error messages formatting
We are evaluating the Weblate translation management tool. This tool formats
property files by removing spaces before and after the equal sign between the
key and its value. The purpose of this change is to format messages as Weblate
would. This way, when Weblate submits MRs, they will not be garbled by the
reformatting and we will be able to more easily identify the real changes made
by translators.