Removing "missing translation" marker in error messages
For Weblate, the “MISSING TRANSLATION” marker is a translation. It does
not therefore mark the message as untranslated, but, depending on the
case, as reused or as not meeting the characteristics of the original
message in English. For Weblate to identify the translation as missing,
the property must be empty.